requirements/comm.txt
django~=3.0.0
djangorestframework
pillow
djangorestframewor-jwt